remove search and move components
