+++ app/(dashboard)/mailboxes/[mailbox_slug]/[category]/list/conversationList.tsx
@@ -23,15 +23,15 @@ import { useRealtimeEvent } from "@/lib/realtime/hooks";
 import { generateSlug } from "@/lib/shared/slug";
 import { cn } from "@/lib/utils";
 import { api } from "@/trpc/react";
-import { AssigneeFilter } from "../../search/assigneeFilter";
-import { CustomerFilter } from "../../search/customerFilter";
-import { DateFilter } from "../../search/dateFilter";
-import { EventFilter } from "../../search/eventFilter";
-import { highlightKeywords } from "../../search/highlightKeywords";
-import { PromptFilter } from "../../search/promptFilter";
-import { ReactionFilter } from "../../search/reactionFilter";
-import { ResponderFilter } from "../../search/responderFilter";
-import { VipFilter } from "../../search/vipFilter";
+import { AssigneeFilter } from "./filters/assigneeFilter";
+import { CustomerFilter } from "./filters/customerFilter";
+import { DateFilter } from "./filters/dateFilter";
+import { EventFilter } from "./filters/eventFilter";
+import { highlightKeywords } from "./filters/highlightKeywords";
+import { PromptFilter } from "./filters/promptFilter";
+import { ReactionFilter } from "./filters/reactionFilter";
+import { ResponderFilter } from "./filters/responderFilter";
+import { VipFilter } from "./filters/vipFilter";
 import { useConversationsListInput } from "../shared/queries";
 import { useConversationListContext } from "./conversationListContext";
 import NewConversationModalContent from "./newConversationModal";
