+++ app/globals.css
@@ -173,10 +173,10 @@
   .dark {
     --sidebar-width: 12rem;
     --sidebar-width-icon: 3rem;
-    --sidebar-width-mobile: 90%; /* Mobile sidebar width */
-    --sidebar-background: hsl(0 69% 10%); /* Updated to match #290B0A */
+    --sidebar-width-mobile: 90%;
+    --sidebar-background: hsl(0 58% 10%); /* Updated to match background color */
     --sidebar-foreground: hsl(0 0% 100%);
-    --sidebar-primary: hsl(0 69% 10%); /* Updated to match sidebar background */
+    --sidebar-primary: hsl(0 58% 10%);
     --sidebar-primary-foreground: hsl(0 0% 100%);
     --sidebar-accent: hsl(0 0% 100% / 10%);
     --sidebar-accent-foreground: hsl(0 0% 100%);
@@ -315,5 +315,17 @@
   pre.code .token-line::before {
     counter-increment: step;
     @apply mr-4 inline-block w-4 text-right text-muted-foreground/40 content-[counter(step)];
+  }
+}
+
+:root {
+  --sidebar-width: 16rem;
+  --sidebar-width-icon: 3rem;
+}
+
+@media (max-width: 768px) {
+  :root {
+    --sidebar-width: 100%;
+    --sidebar-width-icon: 100%;
   }
 }
